[#2449] Support caching of credentials in clients
The JDBC registry now includes a max-age cache directive in its
responses for hashed password and X.509 typed credentials. The max-age
value is configured using the "credentialsTtl" property.

Fixes #2449

+++ services/device-registry-base/src/main/java/org/eclipse/hono/deviceregistry/util/DeviceRegistryUtils.java
@@ -143,7 +143,7 @@ public final class DeviceRegistryUtils {
      *                    returned by the service's operations may be cached for.
      * @return the cache directive corresponding to the given max age for the cache.
      */
-    public static CacheDirective getCacheDirective(final int cacheMaxAge) {
+    public static CacheDirective getCacheDirective(final long cacheMaxAge) {
         if (cacheMaxAge > 0) {
             return CacheDirective.maxAgeDirective(cacheMaxAge);
         } else {
